+++ packages/contracts/test/plugins/governance/majority-voting/addresslist/addresslist-voting-constants.ts
@@ -1,7 +1,0 @@
-import {ethers} from 'hardhat';
-
-export const ADDRESSLIST_VOTING_INTERFACE = new ethers.utils.Interface([
-  'function initialize(address,tuple(uint8,uint32,uint32,uint64,uint256),address[])',
-  'function addAddresses(address[])',
-  'function removeAddresses(address[])',
-]);